iommufd: Add basic skeleton based on liveupdate_file_handler
No functionality is implemented in this commit. Just registering and
unregistering of the struct liveupdate_file_handler for iommufd.

All operations are stubs returning either error or no-op.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>
Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommufd-lu: Implement ioctl to let userspace mark an HWPT to be preserved

Userspace provides a token, which will then be used at restore to
identify this HWPT.

A skeleton for te restore-side ioctl is also added.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>