iommufd: Implement Liveupdate retrieve functionality
Implement LUO retrieve callback to allow retrieval of preserved
iommufds. A retrieved iommufd cannot go through finish until the
preserved HWPTs are restored and all the underlying domain attachments
are replaced with new ones.

The user can restored HWPTs using the new restore hwpt ioctl using the
same token that was used when the HWPT was preserved. The restored hwpt
is associated the underlying restored domain. The restored HWPT is
immutable and it can only destroy or replaced.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>
Signed-off-by: Samiullah Khawaja <skhawaja@google.com>