iommufd-lu: Persist iommu hardware pagetables for live update
The caller is expected to mark each HWPT to be preserved with an
ioctl call, with a token that will be used in restore. At preserve time,
each HWPT's domain is then called with iommu_domain_preserve to get
their iommu token and the token is saved.

On restore, each preserved HWPT is expected to be restored with
another ioctl call, with the same token previously provided to
set_preserved. This HWPT will be recreated without a parent IOAS,
and its domain recreated with iommu_domain_restore. The caller
is expected to later swap the old restored attachments with newly
created HWPTs, though normal means such as VFIO_DEVICE_ATTACH_IOMMUFD_PT.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>