iommufd-lu: Persist iommu hardware pagetables for live update
The caller is expected to mark each HWPT to be preserved with an ioctl
call, with a token that will be used in restore. At preserve time, each
HWPT's domain is then called with iommu_domain_preserve to preserve the
iommu domain.

On restore, each preserved HWPT is expected to be restored with another
ioctl call, This HWPT will be recreated without a parent IOAS, and its
domain recreated with iommu_domain_restore. The caller is expected to
later swap the old restored attachments with newly created HWPTs through
normal means such as VFIO_DEVICE_ATTACH_IOMMUFD_PT.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>
Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommufd: Use the iommu_domain_preserve/unpreserve APIs

Remove the stub implementation of iommufd_domain_preserve and unpreserve
APIs and use the APIs exported by the IOMMU core.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommufd-lu: Make iopt mappings immutable after hwpt is preserved

The map and unmap ioctls operates directly on iopt, which contains
an array of domains, while each hwpt contains only one domain. The
logic then becomes that mapping and nmapping is proibited if
any of the domains in an iopt belongs to a preserved hwpt. However,
tracing to the hwpt through the domain is a lot more tedious than
tracing through the ioas, so if an hwpt is preserved,
hwpt->ioas->iopt is made immutable.

When undoing this (making the iopts mutable again), there's never
a need to make some iopts mutable and some kept immutable, since
the undo only happen on unpreserve and error path of preserve.
Simply iterate all the ioas and clear the immutability flag on all
their iopts.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>

iommufd-lu: Only allow sealed file type iopt_pages

The HWPTs containing dma mappings backed by unpreserved memory should
not be preserved. During preservation check if the mappings contained in
the HWPT being preserved are only file based and all the files are
preserved.

The memfd file preservation check is not enough when preserving iommufd.
The memfd might have shrinked between the mapping and memfd
preservation. This means that if it shrinked some pages that are right
now pinned due to iommu mappings are not preserved with the memfd. Only
allow iommufd preservation when all the iopt_pages are file backed and
the memory file was seal sealed during mapping. This guarantees that all
the pages that were backing memfd when it was mapped are preserved.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>