iommu/vt-d: Restore iommu root_table and context on live update
During boot if the live update state is updated then the iommu live
update state needs to be checked to see if the state of any iommu
hardware unit was persisted before live update. If there is preserved
state available for an iommu hardware unit then restore the root_table
and the iommu context from preserved state.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>