iommu: Add iommu_domain ops to preserve, unpreserve and restore
These domain ops can be implemented by the iommu drivers if they support
iommu domain preservation across liveupdate.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add APIs to preserve/unpreserve iommu domains

These APIs will be used by iommufd LU handler to preserve and unpreserve
a domain for live update.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add API to preserve/unpreserve a device

iommu_preserve_device/iommu_unpreserve_device can be used to
preserve/unpreserve a device for liveupdate. During device preservation
the state of the associated IOMMU is also preserved. The device can only
be preseved if the attached iommu domain is preserved and the assocated
iommu supports preservation.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add APIs to get preserved state of a device

The preserved state of the device needs to be fetched at various places
during liveupdate. The added API can also be used to check if a device
is preserved or not. The API is only used during shutdown and after
liveupdate so no locking needed.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add an API get the preserved state of an IOMMU

After liveupdate kexec during boot, the state of the preserved IOMMU
needs to be restored. Since the state needs to be restored by the IOMMU
drivers during initialization/registration with the IOMMU core, add an
API that can be used by the IOMMU drivers to fetch the preserved state.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add helper APIs to fetch preserved device state

Add two APIs to fetch state of the preserved devices. An API to iterate
through state of all preserved devices and another API to fetch the
state of a single preserved device. Note that these APIs only fetch the
preserved state from the previous kernel.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: Add API to keep track of iommu domain attachments

Keep track of iommu domain attachments by incrementing and decremeting a
counter when a device is attached and detached to a domain.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>