iommupt: Implement preserve/unpreserve/restore callbacks
Implement the iommu domain ops for presevation, unpresevation and
restoration of iommu domains for liveupdate. Use the existing page
walker to preserve the ioptdesc of the top_table and the lower tables.

Preserve top_level, VASZ and FEAT Sign Extended to restore the domain in
the next kernel. The restored domain is made immutable and can only be
freed.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>