vfio/pci: Preserve the iommufd state of the vfio cdev
If the vfio cdev is attached to an iommufd, preserve the state of the
attached iommufd also. Basically preserve the iommu specific state of
the device and also the attach iommu HW unit.

Once the device and its iommufd attachment is preserved, it cannot be
detached or attached to another IOAS until it is unpreserved.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>