iommu/vt-d: restore state of the preserved IOMMU
During boot fetch the preserved state of IOMMU unit and if found then
restore the state. Reuse the root_table that was preserved in the
previous kernel.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu/vt-d: reclaim domain ids of the preserved devices

During IOMMU unit state restore, reclaim the domain ids of the preserved
devices so these are not acquired by another device.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu/vt-d: reuse the preserved domain id for preserved devices

Preserved devices have their domain ids preserved by the previous
kernel. During restore and reattach in the next kernel the domain ids
are already reclaimed so reuse the presered domain id.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>

iommu: restore preserved domain and reattach

During boot the preserved iommu domains need to be recreated and
reattached to the preserved devices. Once the device is reattached to
the preserved domain, the preserved_state set in the dev_iommu can be
unset.

Signed-off-by: Samiullah Khawaja <skhawaja@google.com>